+++ modules/outputs/lib.nix
@@ -50,6 +50,7 @@ in {
       bootPolicy = builtins.hashFile "sha256" ../../nix/automation/boot-media.sh;
     };
     evidence = {
+      toplevel = inputs.self.nixosConfigurations.nixoa.config.system.build.toplevel.outPath;
       sbomnix = inputs.self.packages.${system}.sbomnix.outPath;
       xenOrchestra = inputs.self.packages.${system}.xen-orchestra-ce.outPath;
       xenOrchestraSupply = inputs.self.packages.${system}.xen-orchestra-supply-protector.outPath;